+++ pages/docs/balance/index.mdx
@@ -10,7 +10,7 @@ import CardGrid from "../../../app/components/base/CardGrid";
 import Card from "../../../app/components/base/Card";
 import ProductHeroBanner from "../../../app/components/base/ProductHeroBanner";
 import FooterCTA from "../../../app/components/base/FooterCTA";
-import { ReactComponent as HeroGraphic } from "../../../assets/images/product-icons-and-heroes/dwolla-balance_icon_48x48.svg";
+import { ReactComponent as HeroGraphic } from "../../../assets/images/product-icons-and-heroes/dwolla-balance-icon-48x48.svg";
 
 <ProductHeroBanner
   HeroGraphic={HeroGraphic}
